+++ src/utils.c
@@ -115,32 +115,6 @@ SEXP rnng_sleep(SEXP msec) {
 
 }
 
-SEXP rnng_random(SEXP n) {
-
-  SEXP vec;
-  R_xlen_t vlen;
-
-  switch (TYPEOF(n)) {
-  case INTSXP:
-  case LGLSXP:
-    vlen = (R_xlen_t) INTEGER(n)[0];
-    break;
-  case REALSXP:
-    vlen = (R_xlen_t) Rf_asInteger(n);
-    break;
-  default:
-    Rf_error("'n' must be integer or coercible to integer");
-  }
-
-  vec = Rf_allocVector(INTSXP, vlen);
-  int *pvec = INTEGER(vec);
-  for (R_xlen_t i = 0; i < vlen; i++)
-    pvec[i] = (int) nng_random();
-
-  return vec;
-
-}
-
 SEXP rnng_url_parse(SEXP url) {
 
   const char *up = CHAR(STRING_ELT(url, 0));